% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/node.R
\name{add_child}
\alias{add_child}
\title{Add children node}
\usage{
add_child(main_tree, x, assign_node)
}
\arguments{
\item{main_tree}{the parent tree to be appended with children node}

\item{x}{xth child}

\item{assign_node}{appended node as child}
}
\value{
reshaped tree with children assigned
}
\description{
Add children node
}
\examples{
data("test_df")
data("exercise_df")
test_node <- data.tree::as.Node(test_df)
test_exercise <- data.tree::as.Node(exercise_df)
add_child(main_tree = test_node, x = 4, assign_node = test_exercise )
print(test_node)
}
