% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/honestRF.R
\docType{methods}
\name{addTrees,honestRF-method}
\alias{addTrees,honestRF-method}
\title{addTrees-honestRF}
\usage{
\S4method{addTrees}{honestRF}(object, ntree)
}
\arguments{
\item{object}{A `honestRF` object.}

\item{ntree}{Number of new trees to add}
}
\value{
A `honestRF` object
}
\description{
Add more trees to the existing forest.
}
