% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/node.R
\name{assign_attr}
\alias{assign_attr}
\title{assign attributes to node; work with fix_items function}
\usage{
assign_attr(node_from, node_to)
}
\arguments{
\item{node_from}{assigned attributes from}

\item{node_to}{assigned attributes to}
}
\value{
a node assigned attributes
}
\description{
assign attributes to node; work with fix_items function
}
\examples{
cell_node1 <- data.tree::Node$new("cell1")
cell_node1$AddChild("A")
cell_node2 <- data.tree::Node$new("cell2")
cell_node2$AddChild("A")
cell_node2$Set(group = c(NA, "A1"))
print(assign_attr(node_from = cell_node1$A, node_to = cell_node2$A), "group")

}
