% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forestry.R
\docType{data}
\name{autoforestry-forestry}
\alias{autoforestry-forestry}
\title{autoforestry-forestry}
\format{An object of class \code{NULL} of length 0.}
\arguments{
\item{sampsize}{The size of total samples to draw for the training data.}

\item{num_iter}{Maximum iterations/epochs per configuration. Default is 1024.}

\item{eta}{Downsampling rate. Default value is 2.}

\item{verbose}{if tuning process in verbose mode}
}
\description{
Autotune a forestry based on the input dataset. The methodology
is based on paper `Hyperband: A Novel Bandit-Based Approach to
Hyperparameter Optimization` by Lisha Li, et al.
}
\keyword{datasets}
