% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chart_forest.R
\name{chart_forest}
\alias{chart_forest}
\title{Create plot of variable influence in a random forest model}
\usage{
chart_forest(sim_data, log_var1 = TRUE)
}
\arguments{
\item{sim_data}{Data returned from \link{simulate_data}}

\item{log_var1}{Log-scale the x axis?}
}
\description{
Specify which variables to compare in the model.
}

