% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_detachments.R
\name{compute_detachments-forestry}
\alias{compute_detachments-forestry}
\alias{compute_detachments}
\title{compute detachment indices}
\usage{
compute_detachments(object, feature.new, detachment.feat, p)
}
\arguments{
\item{object}{A `forestry` object.}

\item{feature.new}{A data frame of testing predictors.}

\item{detachment.feat}{A string denoting the feature for computing detachment indices with respect to.}

\item{p}{A positive real number determining the norm p-norm used.}
}
\value{
A vector detachment indices.
}
\description{
return detachment ditances of selected test observations.
}
\examples{
# Set seed for reproductivity
set.seed(292313)

# Use Iris Data
test_idx <- sample(nrow(iris), 10)
x_train <- iris[-test_idx, -1]
y_train <- iris[-test_idx, 1]
x_test <- iris[test_idx, -1]

rf <- forestry(x = x_train, y = y_train)
predictions <- predict(rf, x_test)

# Compute the l-2 detachment indices in the "Petal.Length" dimension
detachments <- compute_detachments(object = rf,
                                   feature.new = x_test,
                                   detachment.feat = "Petal.Length",
                                   p = 2)
}
