% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forestry.R
\docType{methods}
\name{CppToR_translator,externalptr-method}
\alias{CppToR_translator,externalptr-method}
\title{CppToR_translator}
\usage{
\S4method{CppToR_translator}{externalptr}(object)
}
\arguments{
\item{object}{A `forestry` object.}
}
\value{
A list of lists. Each sublist contains the information to span a
  tree.
}
\description{
Add more trees to the existing forest.
}
