% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forestry.R
\name{CppToR_translator}
\alias{CppToR_translator}
\title{Cpp to R translator}
\usage{
CppToR_translator(object)
}
\arguments{
\item{object}{A `forestry` object.}
}
\description{
Translates the forest to a list which can then be used with the
  RToCPP_translator to create an CPP forest object again
}
