% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/node.R
\name{create_nodes}
\alias{create_nodes}
\title{create a tree with assigned name, children and fields}
\usage{
create_nodes(tree_name, add_children_count, ...)
}
\arguments{
\item{tree_name}{assign name of tree}

\item{add_children_count}{assign number of chidren to this tree}

\item{...}{parameters that will be passed as fields of this tree}
}
\value{
a tree with assigned name, children and fields
}
\description{
create a tree with assigned name, children and fields
}
\examples{
create_nodes(tree_name = "tree1", add_children_count = 3, class = c("A", "B", "C"))
}
