% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/node.R
\name{create_tree}
\alias{create_tree}
\title{create tree appended with each element of input list as a child}
\usage{
create_tree(input_list, node_name)
}
\arguments{
\item{input_list}{input list to be made for a tree}

\item{node_name}{name of the tree}
}
\value{
a tree with each item of the list as each child
}
\description{
create tree appended with each element of input list as a child
}
