% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/node.R
\name{cumsum_by_level}
\alias{cumsum_by_level}
\title{calculate cumsum for input level}
\usage{
cumsum_by_level(input_tree, level_num, attri_name)
}
\arguments{
\item{input_tree}{input tree}

\item{level_num}{level of tree for cumsum}

\item{attri_name}{name of this cummulative count field}
}
\value{
tree with calculated cumsum for input level
}
\description{
calculate cumsum for input level
}
