% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/day_of_year.R
\name{day_of_year}
\alias{day_of_year}
\title{day of the year}
\usage{
day_of_year(x)
}
\arguments{
\item{x}{a Date, POSIXct or POSIXlt object.}
}
\value{
return the day of the year as numeric
}
\description{
day of the year
}
