% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate_detachments.R
\name{evaluate_detachments-forestry}
\alias{evaluate_detachments-forestry}
\alias{evaluate_detachments}
\title{compute detachment indices' percentiles}
\usage{
evaluate_detachments(object, feature.new, feat.name, p = 1,
  verbose = TRUE, num.CV = 10)
}
\arguments{
\item{object}{A `forestry` object.}

\item{feature.new}{A data frame of testing predictors.}

\item{feat.name}{a list of features for computing the levels with respect to.}

\item{verbose}{Print out the steps in the algorithm.}

\item{num.CV}{Number of folds in the CV to compute the detachements}
}
\value{
A data frame of quantiles of in response variable conditional on the
test observations.
}
\description{
Computes and returns the detachments percentiles of new observations
in specified dimensions.
}
\details{
compute detachment indices' percentiles
}
\examples{
# Set seed for reproductivity
set.seed(292313)
# Use Iris Data
test_idx <- sample(nrow(iris), 10)

# Select, for example, sepal length as the response variable
index <- which(colnames(iris) == "Sepal.Length")
x_train <- iris[-test_idx, -index]
y_train <- iris[-test_idx, index]
x_test <- iris[test_idx, -index]
rf <- forestry(x = x_train, y = y_train)

features <- c("Sepal.Width", "Petal.Length", "Petal.Width", "Species")

# Evaluate the test observations' detachment indices
trust <- evaluate_detachments(object = rf,
                              feature.new = x_test,
                              feature = features,
                              p = 1)
}
