% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate_lp.R
\name{evaluate_lp-forestry}
\alias{evaluate_lp-forestry}
\alias{evaluate_lp}
\title{compute trust levels}
\usage{
evaluate_lp(object, feature.new, feature, p = 1)
}
\arguments{
\item{object}{A `forestry` object.}

\item{feature.new}{A data frame of testing predictors.}

\item{feature}{a list of features for computing the levels with respect to.}
}
\value{
A data frame of quantiles of in response variable conditional on the
test observations.
}
\description{
Computes and returns the lp levels of new observations in
specified dimensions.
}
\examples{
# Set seed for reproductivity
set.seed(292313)
# Use Iris Data
test_idx <- sample(nrow(iris), 10)

# Select, for example, sepal sength as the response variable
index <- which(colnames(iris) == "Sepal.Length")
x_train <- iris[-test_idx, -index]
y_train <- iris[-test_idx, index]
x_test <- iris[test_idx, -index]
rf <- forestry(x = x_train, y = y_train)

features <- c("Sepal.Width", "Petal.Length", "Petal.Width", "Species")

# Evaluate the test observations' lp distances
trust <- evaluate_lp(object = rf,
                     feature.new = x_test,
                     feature = features,
                     p = 1)
#' @export
}
