% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-exercise_df.R
\docType{data}
\name{exercise_df}
\alias{exercise_df}
\title{Anonymized sample exercise data}
\format{a data frame ready to convert to a tree}
\usage{
data(exercise_df)
}
\description{
Anonymized sample exercise data
}
\examples{
data(exercise_df)
}
\author{
Jiena Gu McLellan, 2020-05-26
}
\keyword{datasets}
