% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/node.R
\name{fill_NA_level}
\alias{fill_NA_level}
\title{fill missing value of a field across a level with 0}
\usage{
fill_NA_level(input_node, field_name, by_level, fill_with = 0)
}
\arguments{
\item{input_node}{input node}

\item{field_name}{field for this operation}

\item{by_level}{across this level}

\item{fill_with}{fill missing value with this value}
}
\value{
node with NA filled for the input field at input level
}
\description{
fill missing value of a field across a level with 0
}
