% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/node.R
\name{fix_items}
\alias{fix_items}
\title{assign certain children nodes and fill NA for empty fields}
\usage{
fix_items(fix_vector, input_node)
}
\arguments{
\item{fix_vector}{children node names to be assigned}

\item{input_node}{the node to be exapnded with children's names}
}
\value{
a node expanded with certain children nodes
}
\description{
assign certain children nodes and fill NA for empty fields
}
\examples{
cell_node2 <- data.tree::Node$new("cell2")
cell_node2$AddChild("B")
cell_node2$AddChild("C")
cell_node2$Set(class = c(NA, "B1", "C1"))
print(cell_node2, "class")
cell_fixed_items <- fix_items(fix_vector = c("A", "B", "C", "D"), input_node = cell_node2)
print(cell_fixed_items, "class")
}
