\docType{package}
\name{forestry-package}
\alias{forestry}
\alias{forestry-package}
\title{Tools for common forestry related analyses.}
\description{
  Tools for common forestry related analyses.
}
\author{
  Marco Bascietto \email{marco.bascietto@ibaf.cnr.it}
}
\references{
  Source code is hosted at GitHub
  (\url{http://mbask.github.com/forestry/})
}
\keyword{package}

