% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantileForest.R
\name{get_conditional_distribution-forestry}
\alias{get_conditional_distribution-forestry}
\alias{get_conditional_distribution}
\title{compute probs}
\usage{
get_conditional_distribution(object, feature.new, vals)
}
\arguments{
\item{object}{A `forestry` object.}

\item{feature.new}{A data frame of testing predictors.}

\item{vals}{values at which values the conditional cdf will be computed}
}
\value{
A data frame of conditional cumulative probabilities
}
\description{
Computes and returns the condidtional distributions.
}
\details{
get_conditional_distribution
}
\examples{
# Set seed for reproductivity
set.seed(292313)

# Use Iris Data
test_idx <- sample(nrow(iris), 10)

# Select, for example, sepal sength as the response variable
index <- which(colnames(iris) == "Sepal.Length")
x_train <- iris[-test_idx, -index]
y_train <- iris[-test_idx, index]
x_test <- iris[test_idx, -index]
rf <- forestry(x = x_train, y = y_train)

vals <- rep(mean(y_train), 10)

# Compute the conditional probabilities associated with values
probs <- get_conditional_distribution(rf, feature.new = x_test, vals)
}
