% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantileForest.R
\name{get_conditional_quantiles-forestry}
\alias{get_conditional_quantiles-forestry}
\alias{get_conditional_quantiles}
\title{compute conditional quantiles}
\usage{
get_conditional_quantiles(object, feature.new, probs)
}
\arguments{
\item{object}{A `forestry` object.}

\item{feature.new}{A data frame of testing predictors.}

\item{probs}{probabilities to be computed at each x}
}
\value{
A data frame of quantiles of in response variable conditional on the
test observations.
}
\description{
Computes and returns the condidtional quantiles for a random
forest.
}
\examples{
#'# Set seed for reproductivity
set.seed(292313)
# Use Iris Data
test_idx <- sample(nrow(iris), 10)

# Select, for example, sepal sength as the response variable
index <- which(colnames(iris) == "Sepal.Length")
x_train <- iris[-test_idx, -index]
y_train <- iris[-test_idx, index]
x_test <- iris[test_idx, -index]
rf <- forestry(x = x_train, y = y_train)

# Compute the  0.25, 0.5, and 0.75 conditional quantiles for sepal length
quants <- get_conditional_quantiles(object = rf,
                                    feature.new = x_test,
                                    quantiles = c(0.25,0.5, 0.75))

# Compare the quantiles to the preditcions
predict(rf, x_test)
}
