% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantileForest.R
\name{get_quantiles-forestry}
\alias{get_quantiles-forestry}
\alias{get_quantiles}
\title{compute quantiles}
\usage{
get_quantiles(object, feature.new, quantiles = c(0.05, 50, 0.95))
}
\arguments{
\item{quantiles}{quantiles to be computed at each x}
}
\value{
A vector of lp distances.
}
\description{
return lp ditances of selected test observations.
}
