% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forestry.R
\docType{methods}
\name{getOOB,forestry-method}
\alias{getOOB,forestry-method}
\title{getOOB-forestry}
\usage{
\S4method{getOOB}{forestry}(object, noWarning = FALSE)
}
\arguments{
\item{object}{A `forestry` object.}

\item{noWarning}{flag to not display warnings}
}
\value{
The OOB error of the forest.
}
\description{
Calculate the out-of-bag error of a given forest.
}
