% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forestry.R
\name{getOOB-forestry}
\alias{getOOB-forestry}
\alias{getOOB}
\title{getOOB-forestry}
\usage{
getOOB(object, noWarning = FALSE)
}
\arguments{
\item{object}{A `forestry` object.}

\item{noWarning}{flag to not display warnings}
}
\description{
Calculate the out-of-bag error of a given forest.
}
