% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forestry.R
\name{getVI}
\alias{getVI}
\title{getVI-forestry}
\usage{
getVI(object, noWarning = FALSE)
}
\arguments{
\item{object}{A trained model object of class "forestry".}

\item{noWarning}{Flag to not display warnings or display warnings.}
}
\description{
Calculate variable importance for `forestry` object as
  introduced by Breiman (2001). Returns a list of percentage increases in
  out-of-bag error when shuffling each feature values and getting
  out-of-bag error.
}
\seealso{
\code{\link{forestry}}
}
