% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chart_forest.R
\name{list_sim_data}
\alias{list_sim_data}
\title{Simulate data from a list of random forests}
\usage{
list_sim_data(rf, class, var1, breaks1 = 50, var2 = NULL, var3 = NULL,
  n_cores = parallel::detectCores(), ...)
}
\arguments{
\item{rf}{random forest}

\item{class}{Which class to plot}

\item{var1}{(Required) Primary variable (preferably continuous)}

\item{breaks1}{How many values of the primary variable should be sampled when
calculating partial dependence?}

\item{var2}{Secondary variable}

\item{n_cores}{How many cores to use when calculating in parallel? Defaults
to all available cores.}
}
\description{
Simulate data from a list of random forests
}

