% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forestry.R
\name{load}
\alias{load}
\title{load RF}
\usage{
load(filename)
}
\arguments{
\item{filename}{a filename in which to store the `forestry` object}
}
\description{
This wrapper function checks the forestry object, makes it
 saveable if needed, and then saves it.
}
