% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forestry.R
\docType{methods}
\name{make_savable-forestry}
\alias{make_savable-forestry}
\alias{make_savable,make_savable-method}
\title{make_savable}
\usage{
\S4method{make_savable}{forestry}(object)
}
\arguments{
\item{object}{an object of class `forestry`}
}
\value{
A list of lists. Each sublist contains the information to span a
  tree.
}
\description{
When a `foresty` object is saved and then reloaded the Cpp
  pointers for the data set and the Cpp forest have to be reconstructed
}
