% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forestry.R
\name{make_savable}
\alias{make_savable}
\alias{make_savable,forestry-method}
\title{make_savable}
\usage{
make_savable(object)
}
\arguments{
\item{object}{an object of class `forestry`}
}
\value{
A list of lists. Each sublist contains the information to span a
  tree.
}
\description{
When a `foresty` object is saved and then reloaded the Cpp
  pointers for the data set and the Cpp forest have to be reconstructed
}
\examples{
set.seed(323652639)
x <- iris[, -1]
y <- iris[, 1]
forest <- forestry(x, y, ntree = 3)
y_pred_before <- predict(forest, x)

forest <- make_savable(forest)
save(forest, file = "forest.Rda")
rm(forest)
load("forest.Rda", verbose = FALSE)
forest <- relinkCPP_prt(forest)

y_pred_after <- predict(forest, x)
testthat::expect_equal(y_pred_before, y_pred_after)
file.remove("forest.Rda")
}

