% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/node.R
\name{pre_get_array}
\alias{pre_get_array}
\title{numericalize children numeric name to convert JSON object to JSON array}
\usage{
pre_get_array(x)
}
\arguments{
\item{x}{input list}
}
\value{
unname numeric names list which is prepared to convert to JSON array
}
\description{
numericalize children numeric name to convert JSON object to JSON array
}
