% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forestry.R
\docType{methods}
\name{predict-forestry}
\alias{predict-forestry}
\alias{predict,forestry-method}
\title{predict-forestry}
\usage{
\S4method{predict}{forestry}(object, feature.new)
}
\arguments{
\item{object}{A `forestry` object.}

\item{feature.new}{A data frame of testing predictors.}
}
\value{
A vector of predicted responses.
}
\description{
Return the prediction from the forest.
}
