% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R_RFNode.R
\docType{data}
\name{RFNode}
\alias{RFNode}
\alias{RFNode}
\alias{RFNode-method}
\title{RFNode Constructor}
\format{An object of class \code{NULL} of length 0.}
\usage{
RFNode(sampleIndex = list(averagingSampleIndex = vector(),
  splittingSampleIndex = vector()), splitFeature = numeric(),
  splitValue = numeric(), child = list())
}
\arguments{
\item{sampleIndex}{A list of index of dataset used in this node and its
children. `sampleIndex` contains two keys `averagingSampleIndex`
and `splittingSampleIndex`. `averagingSampleIndex` is used to generate
aggregated prediction for the node. `splittingSampleIndex` is used for
`honestRF` which stores the splitting data when creating the tree. In
default, `splittingSampleIndex` is the same as `averagingSampleIndex`.}

\item{splitFeature}{Name of the feature that is used for splitting in this
node.}

\item{splitValue}{The value that is used for splitting in this node.}

\item{child}{If the node is not a leaf node, the `child` will be a list of
two `RFNode` objects. If it is a leaf node, the `child` will be `NULL`.}
}
\value{
a `RFNode` object
}
\description{
A constructor to create a RFNode

basic node creator
}
\keyword{datasets}
