% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R_RFTree.R
\name{RFTree}
\alias{RFTree}
\alias{RFTree,RFTree-method}
\title{RFTree-Constructor}
\usage{
RFTree(x, y, se = NULL, mtry = max(floor(ncol(x)/3), 1), nodesize = 5,
  sampleIndex = 1:length(y), splitrule = "variance",
  categoricalFeatureCols = list())
}
\arguments{
\item{x}{A data frame of all training predictors.}

\item{y}{A vector of all training responses.}

\item{se}{#TODO what is SE ???????}

\item{mtry}{The number of variables randomly selected at each split point.
The default value is set to be one third of total number of features of the
training data.}

\item{nodesize}{The minimum observations contained in terminal nodes. The
default value is 5.}

\item{sampleIndex}{A list of the index of observations that are used as
averaging dataset. The index are based on the original dataset `x` and `y`
from forest. Essentially, `x[sampleIndex]` generates the whole splitting
dataset.}

\item{splitrule}{A string to specify how to find the best split among all
candidate feature values. The current version only supports `variance` which
minimizes the overall MSE after splitting. The default value is `variance`.}

\item{categoricalFeatureCols}{A list of index for all categorical data. Used
for trees to detect categorical columns.}
}
\value{
a `RFTree` object
}
\description{
The constructor for the RFTree object
}
