% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forestry.R
\name{saveForestry}
\alias{saveForestry}
\title{save RF}
\usage{
saveForestry(object, filename)
}
\arguments{
\item{object}{an object of class `forestry`}

\item{file}{a filename in which to store the `forestry` object}
}
\description{
This wrapper function checks the forestry object, makes it
 saveable if needed, and then saves it.
}
