% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R_RFTree.R
\docType{methods}
\name{showTree-RFTree}
\alias{showTree-RFTree}
\alias{showTree,RFTree-method}
\alias{showTree}
\title{showTree-RFTree}
\usage{
showTree(object)

\S4method{showTree}{RFTree}(object)
}
\arguments{
\item{object}{A `RFTree`` object}
}
\value{
No return value
}
\description{
Print the entire tree starting from the root node
}
