% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chart_forest.R
\name{simulate_data}
\alias{simulate_data}
\title{Simulate data for partial dependence plotting}
\usage{
simulate_data(rf, d, class, var1, breaks1 = 50, var2 = NULL, breaks2 = 50,
  var3 = NULL, n_cores = parallel::detectCores(), ...)
}
\arguments{
\item{rf}{random forest}

\item{d}{The data frame used to train the original forest}

\item{class}{Which class to plot}

\item{var1}{(Required) Primary variable (preferably continuous)}

\item{breaks1}{How many values of the primary variable should be sampled when
calculating partial dependence?}

\item{var2}{Secondary variable}

\item{breaks2}{If \code{var2} is numeric, in to how many categories should it
be cut?}

\item{n_cores}{How many cores to use when calculating in parallel? Defaults
to all available cores.}

\item{shiny_session}{If a Shiny \link[shiny]{session} object is passed, a
progress bar will be displayed while simulating new data.}
}
\description{
Simulate data for partial dependence plotting
}
