% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-test_df.R
\docType{data}
\name{test_df}
\alias{test_df}
\title{Anonymized sample data}
\format{a data frame ready to convert to a tree}
\usage{
data(test_df)
}
\description{
Anonymized sample data
}
\examples{
data(test_df)
}
\author{
Jiena Gu McLellan, 2020-05-26
}
\keyword{datasets}
