% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forestry.R
\name{testing_data_checker-forestry}
\alias{testing_data_checker}
\alias{testing_data_checker-forestry}
\title{Test data check}
\usage{
testing_data_checker(feature.new)
}
\arguments{
\item{feature.new}{A data frame of testing predictors.}
}
\description{
Check the testing data to do prediction
}

