\name{tpBasalArea}
\alias{tpBasalArea}
\title{Estimate the plot basal area in a traditional plot}
\usage{
  tpBasalArea(treeG)
}
\arguments{
  \item{treeG}{a vector of numeric values od trees basal
  area}
}
\value{
  a numeric estimate for G in the same unit of measure of
  \code{treeG}
}
\description{
  The stimate is simply the sum of trees basal area
}
\author{
  Marco Bascietto \email{marco.bascietto@ibaf.cnr.it}
}
\seealso{
  Other structuralAnalysis: \code{\link{pfExpand}},
  \code{\link{rfBasalArea}}, \code{\link{rfTreeDensity}},
  \code{\link{ttBasalArea}}, \code{\link{ttMeanDbh}}
}

