% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vistree.R
\name{vistree-forestry}
\alias{vistree-forestry}
\alias{vistree}
\title{visualize a tree}
\usage{
vistree(object, tree.id = 1)
}
\arguments{
\item{object}{A forestry object.}

\item{tree.id}{Specifies the tree number that should be visulaized.}
}
\description{
visulizes a tree in the forest.
}
\examples{
set.seed(292315)
library(forestry)
test_idx <- sample(nrow(iris), 3)
x_train <- iris[-test_idx, -1]
y_train <- iris[-test_idx, 1]
x_test <- iris[test_idx, -1]

rf <- forestry(x = x_train, y = y_train)

vistree(rf)
}
