% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forestry.R
\name{addTrees}
\alias{addTrees}
\title{addTrees-forestry}
\usage{
addTrees(object, ntree)
}
\arguments{
\item{object}{A trained model object of class "forestry".}

\item{ntree}{Number of new trees to add.}
}
\value{
A trained model object of class "forestry".
}
\description{
Add more trees to an existing forest.
}
