% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/node.R
\name{assign_attr}
\alias{assign_attr}
\title{assign attributes to node; work with fix_items function}
\usage{
assign_attr(node_from, node_to)
}
\arguments{
\item{node_from}{assigned attributes from}

\item{node_to}{assigned attributes to}
}
\value{
a node assigned attributes
}
\description{
assign attributes to node; work with fix_items function
}
