% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chart_forest.R
\name{chart_forest}
\alias{chart_forest}
\title{Create plot of variable influence in a random forest model}
\usage{
chart_forest(rf, data = NULL, var1, var2 = NULL, var3 = NULL,
  log_var1 = TRUE)
}
\arguments{
\item{rf}{random forest}

\item{data}{Data (if not already in the environment of the random forest)}

\item{var1}{(Required) Primary variable (preferably continuous)}

\item{var2}{Secondary variable}

\item{var3}{Teritary variable}

\item{log_var1}{Log-scale the x axis?}
}
\description{
Specify which variables to compare in the model.
}

