% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_rf_lp.R
\name{compute_lp-forestry}
\alias{compute_lp-forestry}
\alias{compute_lp}
\title{compute lp distances}
\usage{
compute_lp(object, test_set, feature, p)
}
\arguments{
\item{object}{A `forestry` object.}

\item{test_set}{A data frame of testing predictors.}

\item{feature}{A string denoting the dimension for computing lp distances.}

\item{p}{A positive real number determining the norm p-norm used.}

\item{...}{additional arguments.}
}
\value{
A vector lp distances.
}
\description{
return lp ditances of selected test observations.
}
