% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forestry.R
\name{CppToR_translator}
\alias{CppToR_translator}
\title{Cpp to R translator}
\usage{
CppToR_translator(object)
}
\value{
A list of lists. Each sublist contains the information to span a
  tree.
}
\description{
Add more trees to the existing forest.
}
