% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/node.R
\name{cumsum_across_level}
\alias{cumsum_across_level}
\title{cumulative calculation}
\usage{
cumsum_across_level(input_node, attri_name, level_num)
}
\arguments{
\item{input_node}{tree}

\item{attri_name}{name of this cummulative count field}

\item{level_num}{calculate cummulative value cross the level}
}
\value{
tree with cummulative count
}
\description{
cumulative calculation
}
