% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/node.R
\name{fix_items}
\alias{fix_items}
\title{assign certain children nodes and fill NA for empty fields}
\usage{
fix_items(fix_vector, input_node)
}
\arguments{
\item{fix_vector}{children node names to be assigned}

\item{input_node}{the node to be exapnded with children's names}
}
\value{
a node expanded with certain children nodes
}
\description{
assign certain children nodes and fill NA for empty fields
}
