% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forestry.R
\name{getOOB-forestry}
\alias{getOOB-forestry}
\alias{getOOB}
\title{getOOB-forestry}
\usage{
getOOB(object, noWarning)
}
\arguments{
\item{object}{A trained model object of class "forestry".}

\item{noWarning}{Flag to not display warnings.}
}
\value{
The out-of-bag error of the forest.
}
\description{
Calculate the out-of-bag error of a given forest.
}
\seealso{
\code{\link{forestry}}
}
