% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forestry.R
\name{getVI}
\alias{getVI}
\title{getVI-forestry}
\usage{
getVI(object, noWarning = FALSE)
}
\arguments{
\item{object}{A `forestry` object.}

\item{noWarning}{flag to not display warnings or display warnings}
}
\description{
Calculate variable importance for `forestry` object as
  introduced in (Breiman 2001). Returns a list of percentage increases in
  OOB error when shuffling each feature values and getting OOB error.
}
\seealso{
\code{\link{forestry}}
}
