% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forestry.R
\name{getVI.forestry}
\alias{getVI.forestry}
\title{getVI-forestry}
\usage{
getVI.forestry(object, noWarning)
}
\arguments{
\item{object}{A `forestry` object.}

\item{noWarning}{flag to not display warnings}
}
\description{
Calculate increase in OOB for each shuffled feature for forest.
}
