% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forestry.R
\name{predict-forestry}
\alias{predict-forestry}
\alias{predict.multilayerForestry}
\title{predict-forestry}
\usage{
\method{predict}{multilayerForestry}(object, feature.new,
  aggregation = "average")
}
\arguments{
\item{object}{A `multilayerorestry` object.}

\item{feature.new}{A data frame of testing predictors.}

\item{aggregation}{How shall the leaf be aggregated. The default is to return
the mean of the leave `average`. Other options are `weightMatrix`.}
}
\value{
A vector of predicted responses.
}
\description{
Return the prediction from the forest.
}
