% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forestry.R
\name{predict-forestry}
\alias{predict-forestry}
\alias{predict.forestry}
\title{predict-forestry}
\usage{
\method{predict}{forestry}(object, feature.new, aggregation = "average",
  localVariableImportance = FALSE, ...)
}
\arguments{
\item{object}{A `forestry` object.}

\item{feature.new}{A data frame of testing predictors.}

\item{aggregation}{How shall the leaf be aggregated. The default is to return
the mean of the leave `average`. Other options are `weightMatrix`.}

\item{localVariableImportance}{Returns a matrix providing local variable
importance for each prediction.}

\item{...}{additional arguments.}
}
\value{
A vector of predicted responses.
}
\description{
Return the prediction from the forest.
}
