% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forestry.R
\docType{methods}
\name{predict-forestry}
\alias{predict-forestry}
\alias{predict,forestry-method}
\title{predict-forestry}
\usage{
\S4method{predict}{forestry}(object, feature.new, aggregation = "average")
}
\arguments{
\item{object}{A `forestry` object.}

\item{feature.new}{A data frame of testing predictors.}

\item{aggregation}{How shall the leaf be aggregated. The default is to return
the mean of the leave `average`. Other options are `weightmatrix`.}
}
\value{
A vector of predicted responses.
}
\description{
Return the prediction from the forest.
}
