\name{rfBasalArea}
\alias{rfBasalArea}
\title{Estimate the population basal area (\eqn{m^2/ha}) from a relascope plot}
\usage{
  rfBasalArea(relascopeBand = 2, treesCount)
}
\arguments{
  \item{relascopeBand}{the relascopic band used in the
  relascopic plot (1, 2, 4), default: 2}

  \item{treesCount}{the number of trees counted in the
  relascopic plot}
}
\value{
  a numeric estimate for G in (\eqn{m^2/ha})
}
\description{
  Estimate the population basal area (\eqn{m^2/ha}) from a
  relascope plot
}
\author{
  Marco Bascietto \email{marco.bascietto@ibaf.cnr.it}
}

