% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survey_forest.R
\name{survey_forest}
\alias{survey_forest}
\title{Start Shiny app to explore variable importance in a randomForest model}
\usage{
survey_forest(rf, data)
}
\arguments{
\item{rf}{A randomForest object.}

\item{data}{Original data used to train \code{rf}.}
}
\value{
A Shiny app.
}
\description{
Start Shiny app to explore variable importance in a randomForest model
}
